pyproject.toml
[project]
name = "cma"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.93.0",
    "gradio>=6.12.0",
    "ipykernel>=7.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
